fixed the car movement
i figured out why the car wouldnt drive,
 // gameObject.GetComponentInChildren<TextMeshPro>().text = view.Owner.NickName;
this line was causing errors in the start function since were not logged into the multiplayer thing so for testing purposes i commented it out and im gonna leave it like that until we need it again just so everyone knows

i got the car moving again, i updated the car destruction function to add this cool flashing color coroutine thing and then made it so after that it makes the car drivable again. i also make a basic obstacle script that when collisions happen it causees the car to do its die animation and then destroys the obstacle just for testing, but its all set up in 2D and the games 3D so its kinda weird im not sure what were gonna do with that but for now you have to use an object with a 2d collider otherwise nothing works. i also found a really weird bug, if you hit something at an angle that would make your car spin and you dont force the car to straighten out the car just keeps a constant spin added to the car so i made the angular velocity like 10000 on the car object and it seemed like it fixed it and doesnt affect the steering at all. i also added decay to the carstats scrip in case we want to control how fast some cars slow down too passively. im gonna push now i might do more later it depends on how my school work goes.